Move Network declaration into header and rename to NetworkBase
